+++ lib/delegate_builders.dart
@@ -3,3 +3,6 @@
 library delegate_builders;
 
 export 'src/build/delegate/tree_builder.dart';
+export 'src/build/delegate/state.dart';
+export 'src/build/delegate/data_state.dart';
+export 'src/build/tree_builder.dart' show InitialData, InitialChild;